Category Modify

+++ src/main/java/com/Backend/controller/CategoryController.java
@@ -6,9 +6,12 @@ import com.Backend.model.Category;
 import com.Backend.model.User;
 import com.Backend.model.request.DeleteByIdRequest;
 import com.Backend.model.request.InsertCategoryRequest;
+import com.Backend.model.request.ModifyCategory;
 import com.Backend.repository.ICatRepo;
+import com.Backend.repository.IPassRepo;
 import com.Backend.repository.IUserRepo;
 import com.Backend.utils.CategoryUtils;
+import com.Backend.utils.PasswordUtils;
 import net.minidev.json.JSONArray;
 import net.minidev.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,9 +20,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletRequest;
-import java.util.ArrayList;
 import java.util.List;
 
+import static com.Backend.utils.CategoryUtils.getSinCategoria;
 import static com.Backend.utils.JsonUtils.peticionCorrecta;
 import static com.Backend.utils.JsonUtils.peticionErronea;
 import static com.Backend.utils.TokenUtils.getUserIdFromRequest;
@@ -31,6 +34,7 @@ public class CategoryController {
     public static final String INSERTAR_CATEGORIA_URL = "/api/categorias/insertar";
     public static final String ELIMINAR_CATEGORIA_URL = "/api/categorias/eliminar";
     public static final String LISTAR_CATEGORIAS_USUARIO_URL = "/api/categorias/listar";
+    public static final String MODIFICAR_CATEGORIAS_USUARIO_URL = "/api/categorias/modificar";
 
     @Autowired
     ICatRepo repoCat;
@@ -38,52 +42,47 @@ public class CategoryController {
     @Autowired
     IUserRepo repoUser;
 
+    @Autowired
+    IPassRepo repoPass;
+
     public User getUserFromRequest(HttpServletRequest request) throws UserNotFoundException{
         Long id = getUserIdFromRequest(request);
-        User usuario = repoUser.findById(id).orElseThrow(() -> new UserNotFoundException(id));
-        return usuario;
+        return repoUser.findById(id).orElseThrow(() -> new UserNotFoundException(id));
     }
 
     @PostMapping(INSERTAR_CATEGORIA_URL)
     public ResponseEntity<JSONObject> insertar(HttpServletRequest request,
-                                               @RequestBody InsertCategoryRequest idcr)
-            throws UserNotFoundException{
-        JSONObject res = new JSONObject();
+                                               @RequestBody InsertCategoryRequest idcr){
         try {
             User usuario = getUserFromRequest(request);
-            if (!idcr.isValid()) {
-                res.put("statusText", "Los campos no pueden quedar vacíos.");
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(res);
-            }
+            if (!idcr.isValid())
+                return peticionErronea("Los campos no pueden quedar vacíos.");
 
             Boolean existsCat = repoCat.existsByUsuarioAndCategoryName(usuario, idcr.getCategoryName());
             if (existsCat) {
-                res.put("statusText", "Ya existe una categoría con ese nombre");
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(res);
+                return peticionErronea("Ya existe una categoría con ese nombre para el usuario.");
             }
             repoCat.save(new Category(idcr.getCategoryName(), usuario));
-            res.put("statusText", "Categoría creada correctamente");
-            return ResponseEntity.status(HttpStatus.OK).body(res);
+            return peticionCorrecta();
         }
         catch(UserNotFoundException e){
-            res.put("statusText", "Usuario no existente");
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(res);
+            return peticionErronea("El usuario no existe.");
         }
     }
 
     @DeleteMapping(ELIMINAR_CATEGORIA_URL)
     public ResponseEntity<JSONObject> eliminar(@RequestBody DeleteByIdRequest del,
-                                               HttpServletRequest request) throws UserNotFoundException, CategoryNotFoundException {
+                                               HttpServletRequest request) throws UserNotFoundException {
 
         User usuario = getUserFromRequest(request);
-
         try {
             Category cat = repoCat.findById(del.getId()).orElseThrow(() -> new CategoryNotFoundException(del.getId()));
-            if (cat.getUsuario().getId().equals(usuario.getId())) {
+            if (cat.getUsuario().getId().equals(usuario.getId()) && !cat.equals(getSinCategoria(repoCat, usuario))) {
+                PasswordUtils.modifyPasswordsAtCategoryDelete(cat, repoPass, repoCat, cat.getUsuario());
                 repoCat.deleteById(cat.getId());
                 return peticionCorrecta();
             } else
-                return peticionErronea("La categoría " + del.getId() + " pertenece a otro usuario.");
+                return peticionErronea("No se ha podido modificar la categoría.");
         }catch(CategoryNotFoundException c){
             return peticionErronea(c.getMessage());
         }
@@ -102,5 +101,22 @@ public class CategoryController {
         return ResponseEntity.status(HttpStatus.OK).body(res);
     }
 
+    @GetMapping(MODIFICAR_CATEGORIAS_USUARIO_URL)
+    public ResponseEntity<JSONObject> modificar(@RequestBody ModifyCategory modCat,
+                                                HttpServletRequest request) throws UserNotFoundException {
+
+        User usuario = getUserFromRequest(request);
+        if(modCat.isValid()) {
+            Category cat = repoCat.findByUsuarioAndId(usuario, modCat.getId());
+            if(cat.equals(getSinCategoria(repoCat,usuario))) {
+                return peticionErronea("No se permite modificar esa categoría.");
+            }
+            else {
+                cat.setCategoryName(modCat.getCategoryName());
+                return peticionCorrecta();
+            }
+        } else
+            return peticionErronea("No se permiten campos nulos.");
+    }
 
 }
